requirements.txt
fastapi
uvicorn
transformers[torch]
requests
Pillow
pydantic
streamlit